(bsr) libs: export tick function in dedicated file
	fixed not-to-spec dependency-cruiser error
