+++ src/libs/tick.ts
@@ -1,0 +1,7 @@
+export function tick(): Promise<void> {
+  return new Promise((resolve) =>
+    setTimeout(() => {
+      resolve()
+    }, 0)
+  )
+}
